[CDEC-333] Remove Pos.Chain.Block.Blockchain and reorganise modules
1) Remove Blockchain class
2) Create Pos.Chain.Block.Header for GenericBlockHeader and friends
3) Create Pos.Chain.Block.Block for GenericBlock and friends
4) Remove orphan instances from Lens and Instances modules
