+++ chain/src/Pos/Chain/Block.hs
@@ -2,14 +2,16 @@ module Pos.Chain.Block
        ( module X
        ) where
 
-import           Pos.Chain.Block.Blockchain as X
+import           Pos.Chain.Block.Block as X
+import           Pos.Chain.Block.ComponentBlock as X
 import           Pos.Chain.Block.Configuration as X
-import           Pos.Chain.Block.Constructors as X
 import           Pos.Chain.Block.Error as X
 import           Pos.Chain.Block.Genesis as X
+import           Pos.Chain.Block.HasPrevBlock as X
+import           Pos.Chain.Block.Header as X
+import           Pos.Chain.Block.IsHeader as X
 import           Pos.Chain.Block.JsonLog as X
 import           Pos.Chain.Block.Logic as X
 import           Pos.Chain.Block.Main as X
 import           Pos.Chain.Block.Slog as X
 import           Pos.Chain.Block.Types as X
-import           Pos.Chain.Block.Union as X
