+++ chain/src/Pos/Chain/Block/Types.hs
@@ -19,8 +19,9 @@ import           Formatting (Format, bprint, build, later, (%))
 import           Serokell.Util.Text (listJson)
 
 import           Pos.Binary.Class (Cons (..), Field (..), deriveSimpleBi)
+import           Pos.Chain.Block.Block (Block)
+import           Pos.Chain.Block.Header (BlockHeader, HasHeaderHash (..))
 import           Pos.Chain.Block.Slog.Types (SlogUndo (..), buildSlogUndo)
-import           Pos.Chain.Block.Union (Block, BlockHeader, HasHeaderHash (..))
 import           Pos.Chain.Delegation (DlgUndo)
 import           Pos.Chain.Txp (TxpUndo)
 import           Pos.Chain.Update (USUndo)
