--- chain/src/Pos/Chain/Block/Union.hs
@@ -1,7 +1,0 @@
-module Pos.Chain.Block.Union
-       ( module Pos.Chain.Block.Union.ComponentBlock
-       , module Pos.Chain.Block.Union.Types
-       ) where
-
-import           Pos.Chain.Block.Union.ComponentBlock
-import           Pos.Chain.Block.Union.Types
