+++ chain/src/Pos/Chain/Delegation/Class.hs
@@ -25,7 +25,7 @@ import qualified Data.Cache.LRU as LRU
 import           Data.Time.Clock (UTCTime)
 import           Serokell.Data.Memory.Units (Byte)
 
-import           Pos.Chain.Block.Union (HeaderHash)
+import           Pos.Chain.Block.Header (HeaderHash)
 import           Pos.Chain.Delegation.HeavyDlgIndex (ProxySKHeavy)
 import           Pos.Chain.Delegation.Types (DlgMemPool)
 import           Pos.Util.Util (HasLens (..))
