+++ chain/src/Pos/Chain/Ssc/Toss/Class.hs
@@ -14,7 +14,7 @@ import           Control.Monad.Except (ExceptT)
 import           Control.Monad.Trans (MonadTrans)
 
 import           Pos.Chain.Genesis as Genesis (Config)
-import           Pos.Chain.Lrc (RichmenStakes)
+import           Pos.Chain.Lrc.Types (RichmenStakes)
 import           Pos.Chain.Ssc.Commitment (SignedCommitment)
 import           Pos.Chain.Ssc.CommitmentsMap (CommitmentsMap)
 import           Pos.Chain.Ssc.Opening (Opening)
@@ -22,7 +22,7 @@ import           Pos.Chain.Ssc.OpeningsMap (OpeningsMap)
 import           Pos.Chain.Ssc.SharesMap (InnerSharesMap, SharesMap)
 import           Pos.Chain.Ssc.VssCertificate (VssCertificate)
 import           Pos.Chain.Ssc.VssCertificatesMap (VssCertificatesMap)
-import           Pos.Chain.Update (BlockVersionData)
+import           Pos.Chain.Update.BlockVersionData (BlockVersionData)
 import           Pos.Core (EpochIndex, EpochOrSlot, StakeholderId)
 import           Pos.Util.Wlog (WithLogger)
