+++ chain/src/Pos/Chain/Txp/Toil/Logic.hs
@@ -37,7 +37,8 @@ import           Pos.Chain.Txp.Tx (Tx (..), TxId, TxOut (..), txOutAddress)
 import           Pos.Chain.Txp.TxAux (TxAux (..), checkTxAux)
 import           Pos.Chain.Txp.TxOutAux (toaOut)
 import           Pos.Chain.Txp.Undo (TxUndo, TxpUndo)
-import           Pos.Chain.Update (BlockVersionData (..), isBootstrapEraBVD)
+import           Pos.Chain.Update.BlockVersionData (BlockVersionData (..),
+                     isBootstrapEraBVD)
 import           Pos.Core (AddrAttributes (..), AddrStakeDistribution (..),
                      Address, EpochIndex, addrAttributesUnwrapped,
                      isRedeemAddress)
