+++ chain/test/Test/Pos/Chain/Block/CborSpec.hs
@@ -9,7 +9,7 @@ import           Universum
 import           Test.Hspec (Spec, describe)
 import           Test.Hspec.QuickCheck (modifyMaxSuccess)
 
-import qualified Pos.Chain.Block as Core
+import qualified Pos.Chain.Block as Block
 
 import           Test.Pos.Binary.Helpers (binaryTest)
 import           Test.Pos.Chain.Block.Arbitrary ()
@@ -17,39 +17,32 @@ import           Test.Pos.Core.Arbitrary ()
 
 spec :: Spec
 spec = describe "Cbor.Bi instances" $ do
-        -- These data types are defined in the 'core' package which suggests that
-        -- these tests should be there, but they depend on type in eg 'update' so
-        -- that is not possible.
-        -- In addition, these types cannot be moved to the 'core package because
-        -- the 'MainBlockHeader' type is used in the 'infra' package and 'block'
-        -- depends on that.
-        -- Would like to disentangle this at some stage.
         describe "Block types defined in the core package" $ do
               modifyMaxSuccess (min 10) $ describe "GenericBlockHeader" $ do
                   describe "GenesisBlockHeader" $ do
-                      binaryTest @Core.GenesisBlockHeader
+                      binaryTest @Block.GenesisBlockHeader
                   describe "MainBlockHeader" $ do
-                      binaryTest @Core.MainBlockHeader
+                      binaryTest @Block.MainBlockHeader
               describe "GenesisBlockchain" $ do
                   describe "BodyProof" $ do
-                      binaryTest @Core.GenesisExtraHeaderData
-                      binaryTest @Core.GenesisExtraBodyData
-                      binaryTest @(Core.BodyProof Core.GenesisBlockchain)
+                      binaryTest @Block.GenesisExtraHeaderData
+                      binaryTest @Block.GenesisExtraBodyData
+                      binaryTest @Block.GenesisProof
                   describe "ConsensusData" $ do
-                      binaryTest @(Core.ConsensusData Core.GenesisBlockchain)
+                      binaryTest @Block.GenesisConsensusData
                   describe "Body" $ do
-                      binaryTest @(Core.Body Core.GenesisBlockchain)
+                      binaryTest @Block.GenesisBody
               describe "MainBlockchain" $ do
                   describe "BodyProof" $ do
-                      binaryTest @(Core.BodyProof Core.MainBlockchain)
+                      binaryTest @Block.MainProof
                   describe "BlockSignature" $ do
-                      binaryTest @Core.BlockSignature
+                      binaryTest @Block.BlockSignature
                   describe "ConsensusData" $ do
-                      binaryTest @(Core.ConsensusData Core.MainBlockchain)
+                      binaryTest @Block.MainConsensusData
                   modifyMaxSuccess (min 10) $ describe "Body" $ do
-                      binaryTest @(Core.Body Core.MainBlockchain)
+                      binaryTest @Block.MainBody
                   describe "MainToSign" $ do
-                      binaryTest @Core.MainToSign
+                      binaryTest @Block.MainToSign
                   describe "Extra data" $ do
-                      binaryTest @Core.MainExtraHeaderData
-                      binaryTest @Core.MainExtraBodyData
+                      binaryTest @Block.MainExtraHeaderData
+                      binaryTest @Block.MainExtraBodyData
